requirements.txt
google-generativeai
playwright
python-dotenv
